split native and bytecode linkable inputs
Summary:native inputs were being passed around as inputs to both the native and bytecode rules, and converted by regex.

this diff instead threads bytecode inputs through to the bytecode rules and native to native (as before). this eliminates the restriction that native and bytecode have the same filename (and in a followup diff, that native even exists at all).

Test Plan: tests

Reviewed By: yiding

fb-gh-sync-id: 1afebf9
shipit-source-id: 1afebf9