Fix buckconfig documentation formatting error caused by // treated as comment
Summary:In my diff
`Perf: ~200x speedup by using Watchman to search for BUCK files instead of crawling filesystem`,
I introduced an error into `buckconfig.soy` which caused a `<code>` markup block to be
unterminated.

The issue is that Soyweb docs use Closure template syntax, which treats `//` preceded by
whitespace as the beginning of a single-line comment:

https://developers.google.com/closure/templates/docs/concepts

That meant this:

     <code>buck build //path/to/...</code>.)

was turned into this:

     <code>buck build

so the `<code>` block was never terminated.

This fixes the error by using a `{literal}` block to ignore the `//` preceded by
whitespace.

Test Plan:`./docs/soyweb-local.sh`, opened `http://localhost:9111` in a web browser,
confirmed `<code>` markup block in `.buckconfig` section was now terminated correctly.

Reviewed By: bolinfest

fb-gh-sync-id: aa9bf50
shipit-source-id: aa9bf50