Make use of type explicit to fix open source autodeps.
Summary:This is a halfway fix for the open source issue reported by shs96c:
https://github.com/facebook/buck/issues/662

We have some classes that we use internally at Facebook:
* `com.facebook.buck.event.ScubaEvent`
* `com.facebook.buck.event.listener.ScubaBuildConfig`
* `com.facebook.buck.event.listener.ScubaBuildListener`
which we use for recording events to an internal system,
[Scuba](https://research.facebook.com/publications/scuba-diving-into-data-at-facebook/).
We exclude these from the public repo so there is no confusion/possibility
that Facebook logs how external users use Buck (we don't!).
However, these files are included when we run `buck autodeps` internally,
which is reflected in the `BUCK.autodeps` files that get checked in.

In practice, these source files are picked up by `glob()` rules in their respective
`BUCK` files, so there is no need to modify the `BUCK` files when we publish
them externally. What this change brings to light is that external users
have been pulling in a few extra deps for `//src/com/facebook/buck/event:event` and
`//src/com/facebook/buck/event/listener:listener`.

In general, additional deps would add some drag to the open source build, but shouldn't
break it. It did in this case because imports in the Scuba classes were compensating
for implicit dependencies elsewhere in the package. This is a common issue,
which is explained in the [Troubleshooting](https://buckbuild.com/command/autodeps.html#troubleshooting)
section of the docs under "Fixing Underspecified Deps."
https://github.com/facebook/buck/commit/820a985220db327bff29c20a411f14f9a8f4f9c3
is an example of how we [thought we] fixed this for the rest of the repo
in anticipation of introducing `buck autodeps`.

The ideal solution is to have a plug-in API for our Scuba integration so
that we can exclude this code from the internal Buck repo entirely.
The current solution has its own drawbacks, though (see the Test Plan section).
In the interim, it might be reasonable to exclude the `BUCK.autodeps` files from
the open source build and require external users to generate them themselves.

Test Plan:Applied this change to the open source version of Buck and ran `buck autodeps` on it.
The result is that there still end up being local changes to the `BUCK.autodeps` files
(but now `./bin/buck build buck` succeeds):

```
$ git diff
 diff --git a/src/com/facebook/buck/event/BUCK.autodeps b/src/com/facebook/buck/event/BUCK.autodeps
index 13a97ec..e7a5e60 100644
 --- a/src/com/facebook/buck/event/BUCK.autodeps
+++ b/src/com/facebook/buck/event/BUCK.autodeps
@@ -1,4 +1,4 @@
-#@# GENERATED FILE: DO NOT MODIFY 563d1548299b9f50d34508fe0cc9f5a05d0514dc #@#
+#@# GENERATED FILE: DO NOT MODIFY 3091e9073709aa561c42c3fe78610c1fbc34e98c #@#
 {
   "event" : [
     "//src/com/facebook/buck/log:api",
@@ -12,7 +12,6 @@
     "//third-party/java/guava:guava",
     "//third-party/java/immutables:processor",
     "//third-party/java/jackson:jackson-annotations",
-    "//third-party/java/jackson:jackson-databind",
     "//third-party/java/jsr:jsr305"
   ]
 }
 diff --git a/src/com/facebook/buck/event/listener/BUCK.autodeps b/src/com/facebook/buck/event/listener/BUCK.autodeps
index db6c047..70b8535 100644
 --- a/src/com/facebook/buck/event/listener/BUCK.autodeps
+++ b/src/com/facebook/buck/event/listener/BUCK.autodeps
@@ -1,7 +1,6 @@
-#@# GENERATED FILE: DO NOT MODIFY 0b2f0a60588d321e3f62fcbcbbc7e6caab3e80fd #@#
+#@# GENERATED FILE: DO NOT MODIFY 2bdebcaeba40b92970b5b7babaa19820e2aeb249 #@#
 {
   "listener" : [
-    "//src/com/facebook/buck/android:steps",
     "//src/com/facebook/buck/artifact_cache:artifact_cache",
     "//src/com/facebook/buck/cli:config",
     "//src/com/facebook/buck/cli:events",
@@ -30,14 +29,11 @@
     "//src/com/facebook/buck/util:constants",
     "//src/com/facebook/buck/util:exceptions",
     "//src/com/facebook/buck/util:io",
-    "//src/com/facebook/buck/util:tristate",
     "//src/com/facebook/buck/util:util",
     "//src/com/facebook/buck/util/concurrent:concurrent",
     "//src/com/facebook/buck/util/environment:environment",
-    "//src/com/facebook/buck/util/environment:platform",
     "//src/com/facebook/buck/util/network:network",
     "//src/com/facebook/buck/util/unit:unit",
-    "//src/com/facebook/buck/util/versioncontrol:versioncontrol",
     "//third-party/java/guava:guava",
     "//third-party/java/jackson:jackson-core",
     "//third-party/java/jackson:jackson-databind",
```

So on the plus side, external users can build again.

The downside is that if an external users wants to put up a pull request,
and is using `buck autodeps` (as we would expect), then the PR would likely
include those changes to `buck autodeps`, which would break the internal
version of Buck.

Reviewed By: Coneko

fb-gh-sync-id: f74771b
shipit-source-id: f74771b