C/C++: fix how we implicitly copy compiler flags to preprocessor flags
Summary:We always apply compiler flags to preprocessing, since in many cases
these case implicit preprocessor macros to be defined, and it's really
difficult/tedious for users to figure out which ones are needed manually.
Previously, we did this in several different places, and it led to
unintentionally apply the compiler flags twice.

This diff does moves this to a single location, so it's easier to do
this correctly.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: 79d2a5a
shipit-source-id: 79d2a5a