fix color diagnostics for pch rules
Summary:Instead of picking from preprocessorCommand or compilerCommand based on whether
the operation is preprocess or not, just use whichever is available, since they
tend to both either have color enabled, or not.

Also pass in the whether the command would have output color, instead of the
command itself, because semantically it's as much info as is needed.

Test Plan: Made an error in prefix header with PCH on, see color diagnostics instead of crashing.

Reviewed By: Coneko

fb-gh-sync-id: b46aada
shipit-source-id: b46aada