Cache the work done by `buck autodeps`.
Summary:Now `buck autodeps` records the provided/required symbols that it has extracted
for a rule in the Buck build cache. This means that incremental runs of `buck autodeps`
are considerably faster because it no longer has to re-parse all of the `.java` files
in the project: it will only re-parse the ones that are part of build rules whose `srcs` have been modified.

For large projects that use a shared build cache, it likely makes sense to run `buck autodeps` in CI
in order to populate the cache.

Test Plan: Added tests: `SymbolsTest`, `JavaLibrarySymbolsFinderTest`, and `JavaSymbolsRuleTest`.

Reviewed By: Coneko

fb-gh-sync-id: 6bc4b18
shipit-source-id: 6bc4b18