Refactor cache invalidation events in DaemonicParserState
Summary:I went to add another counter for daemon cache invalidation events
in `DaemonicParserState`, when I realized the existing code wasn't
correctly handling the case when the cache was empty to begin with.

I had previously tried to handle this by making the environment
variable cache an `Optional`, but this only handled the first startup
case.

If we invalidate the cache but nothing was in the cache in the first
place, there's no need to log. This diff refactors the cache invalidation
logic in `DaemonicParserState` to:

1) Remember whether or not there was anything in the cache when we invalidate it, and
2) Only bump the cache invalidation counters if the cache was non-empty when invalidated.

Test Plan:`FOO=bar buck build buck; FOO=baz buck build buck`, confirmed in buck-out/log/buck-0.log
that change was detected, cache was invalidated, and counters were logged.
Confirmed in buck-out/log/buck-1.log that empty cache being invalidated was not treated as
an event to be logged.

Reviewed By: Coneko

fb-gh-sync-id: a3de386
shipit-source-id: a3de386