Make "as" point to the C compiler frontend
Summary:Until now there was no single place in our codebase that would read
the location of the `as` tool in the `CxxPlatform`. Instead we just
used the C compiler to compile assembly. Since having an unused
field is pointless, there are two ways to fix it: either remove it
or put the value that is actually used in there. I have decided for
the latter, partially because we already have `asflags` that
contain flags meant for the C compiler when compiling assembly,
partially because for example for iOS it is not appropriate to use
the `ClangPreprocessor` and `ClangCompiler` for assembly code
because when `clang` compiles assembly, it does not support many of
the features it supports when it compiles C or C++ code.

Test Plan: `buck test`, and compiling some assembly with Buck.

Reviewed By: Coneko, andrewjcg

fb-gh-sync-id: 08d2fc4
shipit-source-id: 08d2fc4