Add a kill switch to parallel parsing.
Summary:If `project.parallel_parsing` == false we still perform marshalling
on a threadpool. This introduces an extra degree of concurrency
that has caused a deadlock in the past. Running the whole thing
on a directExecutor makes the code behave in a more predictable way
(and closer to the original single-threaded implementation)

Test Plan: CI, run `buck targets` with parallel parsing on and off.

Reviewed By: k21

fb-gh-sync-id: 0347bf4
shipit-source-id: 0347bf4