Detect whether compiler is Clang
Summary:Replace the current logic that guesses which compiler we are using
based on the host platform with a proper detection of the Clang
compiler based on its version information. To avoid running the
compiler eagerly, replace the fields of `CxxPlatform` with
suppliers that run the compiler and figure out what kind of
compiler it is when the field is first accessed.

Test Plan:`buck test`, check that running clang on linux uses its color
diagnostics.

Reviewed By: Coneko

fb-gh-sync-id: 8b0716f
shipit-source-id: 8b0716f