Change ParserLeaseVendor to be ProjectBuildFileParserPool
Summary:This alters the ParserLeaseVendor to not depend on the Future
callback to keep track of whether a parser is in use or not. The problem
with the callback approach was the fact that canceling the future would run the
callback immediately, regardless of whether we were done using the parser
or not.
This also prevents the parser from leaking out by calling the
getAllRulesAndMetaRules method directly (which is what motivated the rename).

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: af2243c
shipit-source-id: af2243c