Fix 2 level caching metadata handling.
Summary:The two-level cache was storing the metadata in the second level (with the
artifact which was only indexed by content hash). This meant that any
RuleKeys that shared the same output hash could end up getting
metadata associated with any of the other RuleKeys in that group.
The fix is to store the metadata along with the RuleKey (in the first
level of the cache).

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: 9e5a601
shipit-source-id: 9e5a601