implemented getPathToOutput and getRuntimeDeps in DBinary and DTest
Summary: DBinary and DTest were missing getPathToOutput and getRuntimeDeps. The former causes the path to the generated files not to be listed when --show-output is passed. The latter causes no files to be written when we get cache hits on D binaries and D tests. This commit fixes both of those issues.

Test Plan:bin/buck test test/...
Built a d_binary with dir cache, removed the buck-out/gen and buck-out/bin directories, and ran the build again. Did the same for a d_test.
Without this change, the buck-out/gen directory is not regenerated, so the binary we were supposed to build is missing. With the change, the binary is present.
Ran buck build --show-output on a d_binary and saw that it prints the path to the binary now.

Reviewed By: andrewjcg

fb-gh-sync-id: 993a23a
shipit-source-id: 993a23a