UncachableBuildRule, for the rules that you just can't cache.
Summary:UncachableBuildRule is a new interface for rules that should not be cached.

This marker will prevent the cache from being read or written to, but will
still avoid building if the build product on disk is current (in both standard
and input-based senses of 'current').

This is different from doing work in post build steps as post build steps are
meant to run on the fetched artifact, and does not wait for dependencies to be
built.

This will be useful for precompiled header rule, which bakes in mtimes and hard
coded paths that are difficult to sanitize, and which generally runs fast
enough such that the sanitizing time is greater than just running the rule.

Test Plan: a unit test to check that caches are not touched.

Reviewed By: andrewjcg

fb-gh-sync-id: e9c4356
shipit-source-id: e9c4356