Add pmd_extensions jar to allow blacklisting APIs
Summary:We'd like to be able to blacklist specific APIs at lint time.

We do have the `forbiddenapis` tool, but it has a ton of APIs built-in
that we're not ready to clean up.

So, this imports the `pmd_extensions` open-source library (Apache 2.0
license) which adds a few very useful PMD lint rules:

http://liveramp.com/engineering/using-pmd-to-blacklist-unsafe-methods/

When I tested it, I had to fix a few issues. I created a branch
here and pushed the fixes, for which I will send pull requests upstream:

https://github.com/bhamiltoncx/pmd_extensions

1) It depended on Apache Commons for a single method
2) It needed to be ported to PMD 5.1 (we use 5.1.2)
3) When blacklisting a method, the alternative suggestion overwrote the
original blacklist description

Test Plan: Added PMD rule in subsequent diff, ran `ant pmd`.

Reviewed By: bolinfest

fb-gh-sync-id: 4c94365
shipit-source-id: 4c94365