Fix defect where a class that is abstract but has @Test annotations w…
Summary:…as attempted to be instantiated and threw java.lang.InstantiationException

If you have an abstract Java class that defines methods with Test annotations, buck attempts to instantiate an object with the class type and fails with a java.lang.InstantiationException.

This patch skips abstract classes when selecting tests to run. We hit this error trying to port ONOS from maven to buck.
Closes https://github.com/facebook/buck/pull/696

Reviewed By: sdwilsh

fb-gh-sync-id: 4e1eabb
shipit-source-id: 4e1eabb