Add ability to insert values in AndroidManifest from BUCK files
Summary:This change adds a ManifestEntries data type to the Android builder.
This data class allows definition of some important manifest entries
in the buck file, in a manner similar to the Gradle DSL.
(See http://tools.android.com/tech-docs/new-build-system/user-guide#TOC-Basic-Build-Customization
for details) on the Gradle DSL.

Included in this change:

- The ManifestEntries Immutable Value declaration with optional fields for
 - minSdkVersion
 - targetSdkVersion
 - versionCode
 - versionName
 - debugMode

- Additions to AaptStep which use the ManifestEntries to generate command line flags
  so that AAPT can insert the appropriate values into the manifest (as well as the flag
  which allows errors with conflicting values to surface)

NOTE: Previous versions of the diff contained a macro to check the keys. I've decided to add checking to the type coercer instead.

- A TypeCoercer for ManifestEntries which can load from a normal dict.
```
default_config = {
  'min_sdk_version': 14,
  'target_sdk_version': 23,
  'version_name': 'Yosemite',
}
```
- An optional parameter for the AndroidBinary rule which takes a list of these
  ManifestEntries and can merge them together.

```
android_binary(
  # Other params here...
  configs = [
    default_config,
    dict(version_code=42)
  ]
)
```

The goal is to allow for re-use of a base config, but also to allow customization
and specialization for different build flavors.

Test Plan:Added unit tests for AaptStep and all added ManifestEntries classes.
Verified that generated APK manifest correctly reflects the contents of the configs.

Reviewed By: marcinkosiba

fb-gh-sync-id: f817dc4
shipit-source-id: f817dc4