Disable support for running instrumentation tests on multiple devices
Summary:The instrumentation tests are going to be run in a similar fashion as
the junit tests. E.g. running as a shell command in a different process
as opposed to running directly from inside the buck process. In that
model supporting running tests on multiple devices all at once is
tricky. I'll add it back if it turns out people really want it.

Note that the implementation would only run against multiple devices if
you explicitly passed the `-all` or `-x` flag to the buck test command
and if you did all the tests would run serially one device after
another. If we want to reincarnate support down the line could probably
at least try to run against the devices in parallel if asked to do so.

Test Plan: buck test + ran an instrumentation test with a newly built buck

Reviewed By: Coneko

fb-gh-sync-id: 30f3913
shipit-source-id: 30f3913