CachingBuildEngine: use a new weighted executor
Summary:Currently, the number of threads determines how rules can be built
locally in parallel, where one building rule occupies a single thread.
However, this 1:1 mapping of rules to build slots doesn't always work
well in builds with a diverse set of rules.  In some cases, one phase
of the build may be filled with light-weight rules which run with
really high-levels or parallelism, whereas another phase might be
dominated by a completely different rule type which uses large amounts
of system resources, and so has a really low-level of parallelism.

This diff adds a new `WeightedListeningExecutor` class which uses a
new `ListeningSemaphore`.  This gates a wrapped `ListeningExecutor`
and allows executing jobs to use a custom weight.  Using this and a
specially configured thread count and default job weight, it should
be possible to use custom rule weights to better fit different types
of builds.

Test Plan: Added unittests.

Reviewed By: marcinkosiba

fb-gh-sync-id: da36c74
shipit-source-id: da36c74