Update `buck autodeps` to generate exported_deps as well as deps.
Summary:This addresses a major pain point with `buck autodeps`. Historically, the
`import` statements in your `srcs` did not specify all of your deps because
you could have chained expressions in your Java code that introduced implicit
dependencies. This required one of the following workarounds:

1. Caller adds the missing dependency to its `deps`.
2. Callee adds the missing dependency to its `exported_deps`.

In general, the latter is more appropriate because it also solves the problem
for the next person.

With this change, we do the latter for the user via `BUCK.autodeps` by figuring
out which of a rule's deps contribute to its non-private API and then include
those as `exported_deps`. This will ultimately allow us to remove many hand-maintained
lists of `exported_deps` for build rules in the Buck codebase. It should also
make it easier to use `autodeps = True` in the rules in the `test/` folder that
have not been converted yet.

This change is going to require all `BUCK.autodeps` files to be rewritten.
I am doing that work in a separate revision (stacked on top) to make this
easier to review, so do not be surprised that there are CI failures.

Test Plan: CI in subsequent revision.

Reviewed By: yiding

fb-gh-sync-id: a4d2edb
shipit-source-id: a4d2edb