Use fully-qualified build targets in exported_deps.
Summary:This is a temporary measure to work around an edge case.
The edge case is now documented in `buck.py`.
It should be addressed properly, though it will be much less
of an issue, in practice, once I get this current stack of
diffs to make changes to `buck autodeps` landed.

Test Plan: CI

Reviewed By: yiding

fb-gh-sync-id: 3a420db
shipit-source-id: 3a420db