Migrate test/ to use autodeps.
Summary:At first, I started this revision by finding `java_test()` and `java_library()` rules
under the `test/` folder and removing the `deps` and `exported_deps` arguments
and replacing them with `autodeps = True`. However, the majority of the `BUCK`
files under `test/` follow the following pattern:

```
java_library(
  name = 'testutil',
  srcs = glob(['*.java'], excludes = ['*Test.java']),
  autodeps = True,
  visibility = [
    '//test/...',
  ],
)

java_test(
  name = <NAME_OF_DIRECTORY>,
  srcs = glob(['*Test.java']),
  resources = glob(['testdata/**'], include_dotfiles=True),
  autodeps = True,
)
```

So to eliminate this boilerplate, I introduced a macro:

    standard_java_test(run_test_separately = False)

that generates the above rules for the `srcs` in the directory.
Of the 90 `BUCK` files in the `test/` directory, 61 of them (or 68%) could
be reduced to a single call to `standard_java_test()`. I suspect
that if we put in the effort, we could get that number to 90%.
This will reduce a bunch of boilerplate in creating `BUCK` files for tests.

For example, here are two files that should be able to use the macro, but cannot
due to small bugs:

```
test/com/facebook/buck/maven/BUCK
test/com/facebook/buck/testutil/integration/BUCK
```

To find other files that are candidates for using the macro, run:

```
find test -name BUCK | xargs grep -l -e '^java_test' | wc -l
```

Finally, note this revision added a `BUCK` file, as it was missing:

    test/com/facebook/buck/cxx/elf/BUCK

Apparently `ElfTest.java` was not being run.

Test Plan: CI

Reviewed By: yiding

fb-gh-sync-id: bbc81b5
shipit-source-id: bbc81b5