SymlinkTree: defer verifying paths until build time
Summary:Previously, we'd verify the path components of a `SymlinkTree` in it's
constructor, which meant we do some up-front work and required having
callers catch and deal with an exception (and in many cases, the caller
just used a try-catch to propagate it as a `RuntimeException`).

This diff moves this verification to be done during the actual build
steps.  This does also mean that other parts of the codebase that may
have been relying on the `SymlinkTree` constructor to validate these
paths may have to do their own explicit validating, which they likely
should have done from the beginning.

Test Plan: refactored test

Reviewed By: yiding

fb-gh-sync-id: edd289b
shipit-source-id: edd289b