Put experimental project cleaner behind a switch
Summary:This diff puts the cleaner (which removes any non-generated files) behind
a switch to avoid causing problems for source control systems on repositories
where other .iml files have been checked in.

This means that, by default, the behaviour will be the same as before the
cleaner was introduced.

Test Plan: - Local testing

fb-gh-sync-id: 919c525
shipit-source-id: 919c525