Implement tracking of unskipped rules
Summary:Track how many rules are either already executed or might be
executed in the future and update the rule count shown in the
console based on that. This way, the number of executed rules at
the end of the build will always match with the displayed total
number of rules.

Test Plan:* New tests for `UnskippedRulesTracker`.
  * Run a build with partial cache misses, check that the number
    of rules goes down as we get cache hits and at the end of the
    build matches the number of executed rules.

Reviewed By: marcinkosiba

fb-gh-sync-id: 338d977
shipit-source-id: 338d977