Add annotation paths to the generated IntelliJ modules.
Summary:Add the path for source code generated by an annotation processor to the modules where the annotations occur.

This adds the path under a separate ContentRoot, which may cause issues older versions of IntelliJ, but the latest version of IJ15 and 2016.1 can cope with this.

To force building of the annotation processor the --process-annotations parameter must be passed to buck project. Without it buck may use cached build artifacts which do not include the source code generated by annotation processors.

Test Plan: - Local testing

Reviewed By: ASCE1885

fb-gh-sync-id: db6fedd
shipit-source-id: db6fedd