Always write out the seeds file for each ProGuard run
Summary:This revision adds the dumping of the ProGaurd seeds set in the same way that the mapping file is dumped.
We have to take care to ensure there is always at least one keep directive in the ProGuard configuration
otherwise ProGuard will exist with an error. However, all Android builds will satisfy this requirement.
To be super paranoid, we could add code to walk all the config files and check to see if any line starts with "-keep"
(modulo white space and comment #) and only then print out the seeds. However, this seems perhaps too
paranoid and expensive.

Test Plan:I have manually tested this by making an FB4A build and observing the generated seeds file:

  $ mybuck build //apps/fb4a:fbandroid_staticdi_dextr_comp_downloadable_str_xzs_xhdpi_armv7_release_optimizedresources_redex_fbsign
  [-] PROCESSING BUCK FILES...FINISHED 57.6s [100%]
  [-] DOWNLOADING... (311.90 KB/S AVG, TOTAL: 132.06 MB, 34034 Artifacts)
  [-] BUILDING...FINISHED 3485.3s [100%] (16672/16672 JOBS, 16672 UPDATED, 99.8% CACHE MISS, 0.2% CACHE ERRORS) Details: http://localhost:54820/trace/c80d6dd1-c886-43bb-882a-f21
  $ find buck-out -name seeds.txt
  buck-out/gen/apps/fb4a/__fbandroid_staticdi_dextr_comp_downloadable_str_xzs_xhdpi_armv7_release#aapt_package__proguard__/.proguard/seeds.txt

Reviewed By: dreiss

fb-gh-sync-id: a9d760e
fbshipit-source-id: a9d760e