Only clean dir cache on close if cache was written to
Summary:Previously this always cleans after every command. This is problematic for
interactive commands like `--help`, as the command wouldn't exit and has to be
ctrl-c'd.

This is a sorta-fix that only cleans the cache if the cache has stuff written
to (and did not have a clean done).

Test Plan:1. Populate the dir cache with gigabytes of stuff.
2. `time buck --help`
3. See that it doesn't stick around after printing the usage string.

Reviewed By: beefon

fb-gh-sync-id: 7fed069
fbshipit-source-id: 7fed069