Ignore some events in HangMonitor
Summary:Since `CountersSnapshotEvent` is sent out periodically every few
seconds, `HangDetector` thought that a progress is made even in
case of a deadlock and never reported an issue. This changes
`HangDetector` to only consider events that are a result of a bit
of work being finished, not events that are sent out periodically.

Test Plan:Add a long sleep to `Build.java`, check that `HangMonitor` reports
threads' stacks.

Reviewed By: marcinkosiba

fb-gh-sync-id: 6c73d1a
fbshipit-source-id: 6c73d1a