Behold ActionGraphAndResolver
Summary:Refactoring the pair <ActionGraph, BuildRuleResolver> to an Immutable object for better readability.
This will also help when we implement ActionGraphCache so we do not have to store pairs of the graph
and the resolver.

The object could also be a immutable tuple but the object improves the readability from the places
that use only one of the fields, graph or resolver.

Test Plan:buck autodeps; buck build buck;

CI

Reviewed By: Coneko

fb-gh-sync-id: 52398ed
fbshipit-source-id: 52398ed