Added package name to buck install event
Summary:I want to add buck install debug to the plugin.
My main problem was the fact that  I needed the installed package name in order to find out to what process to attach, so this adds the installed package name to the install finished event.

Test Plan:The package name received in plugin is valid.
It passes all buck tests.
'buck install' still works

Reviewed By: alsutton

fb-gh-sync-id: 95117f0
fbshipit-source-id: 95117f0