New ContentAgnosticRuleKeyBuilderFactory
Summary:This diff is related with the testing period of ActionGraphCache.

During the initial testing period of the ActionGraphCache we will still create the ActionGraph each time. If we should return the cached version, we will compare the RuleKeys of the 2 graphs. If the change is only in the file contents the RuleKeys should be the same. For every other change they should change. Right now the RuleKeyBuilderFactory creates a hash from the file's contents and takes it into account so the comparison is not possible.

ContentAgnosticRuleKeyBuilderFactory creates RuleKeys that only take into account the filename etc and not the file contents. This class will be used to make sure that the comparison happens correctly and that the RuleKeys change with the same principle we should/shouldn't dump the ActionGraph.

Test Plan:buck build buck;

Unit-tests:
- If you change the file contents of a file the RuleKeys should stay the same.
- If you "rename" a file the RuleKey should change.

Reviewed By: marcinkosiba

fb-gh-sync-id: f1af316
fbshipit-source-id: f1af316