Fix logic for transitive classpath entries in android_binary rule
Summary:The current implementation of getTransitiveClasspathEntries for android_binary
rules suffers from the following problem:

Due to the different ways that AndroidBinaryGraphEnhancer and
Classpaths.getClasspathEntries traverse the dependency graph there are cases
that jars end up in the apk but are not visible in the result of the
classpath. The most common example are android_library rules as deps
of android_resource rules. AndroidBinaryGraphEnhancer traverses
android_resource deps and finds AndroidPackageable android_library rules
behind android_resource rules while Classpaths.getClasspathEntries doesn't
look further after encountering an android_resource.

This diff fixes the problem by reworking the implementation of
getTransitiveClasspathEntries and getTransitiveClasspathDeps for AndroidBinary
Basically the AndroidBinaryGraphEnhancer returns the appropriate SourcePath
collection in AndroidGraphEnhancementResult.getClasspathEntriesToDex
instead of regular Paths and AndroidBinary returns the appropriate collections
after transforming the SourcePaths to BuildRules and invoking the Classpaths
methods on those.

Test Plan:ant test
Confirm that classpath on an android_binary rule contains hidden modules
buck audit classpath on a repro case
(basically create a android_library rule that is only ever used in the binary
as a dep to an android_resource rule)

Reviewed By: marcinkosiba

fb-gh-sync-id: 3b0dbd8
fbshipit-source-id: 3b0dbd8