Cache the ActionGraph
Summary:Currently we cache the TargetNodes of the TargetGraph but we do not cache the ActionGraph. A common building pattern is to edit files and build again. In this case we re-create the ActionGraph each time and it takes time. One optimization would be to cache the ActionGraph and re-use it in case there are no file additions-deletions that would invalidate the TargetNodes anyway.

This diff creates ActionGraphCache. We cache the last ActionGraph by using the TargetGraph as a key. To enable TargetGraphs we implement equals & hashcode. If the cached and new versions of the TargetGraph match then we should return the cached version. For testing purposes we still create a new ActionGraph each time and using ContentAgnosticBuildRuleKeyBuilderFactory we compare the two ActionGraphs through their RuleKeys. The reason we have a different RuleKeyBuilderFactory is because the alterations to the contents of a file do not change an ActionGraph but alter the default RuleKeys. The comparison of the ActionGraphs adds some overhead during the testing period but it will be removed when we are sure we return what is expected. The measurement of cache misses, hits and mismatch between new and cached ActionGraphs is being done by Counters. Note that the Counters are not being registered when we run with NO_BUCKD=1; since then we expect 100% cache misses.

Test Plan:- Make the changes; buck autodeps; buck build buck;
- Unit tests: Make sure the returned ActionGraph is the expected one and cache hits and misses are counted correctly.
- Memory profiling: I used a profiler to check memory consumption during building an app with a big graph, some increase but nothing major or weird.

Reviewed By: marcinkosiba

fb-gh-sync-id: 1c5577e
fbshipit-source-id: 1c5577e