Handle cross-cell information in `buck audit input`
Summary:There are two somewhat reasonable ways to approach this:
1. Use the cell path and relative paths.
2. Use the cell name as defined by the config

The first has the advantage of giving us the path we want and roughly "works"
The second has the advantage of giving us Buck-understable targets

This mostly depends on who depends on buck audit and how things might break. Also the second depends on a refactor to pass cells through build targets and the build target cache.

Test Plan: unit tests

Reviewed By: Coneko

fb-gh-sync-id: 5277f46
fbshipit-source-id: 5277f46