Fix flaky build engine test
Summary:The `BuildRuleEventTests` ordered rule events by timestamp to verify
begin/end pairs ran on the same thread.  However, it was possible that
a end and begin event had the same timestamp which caused the tests
to pair up two begin events which would run on different threads.

This diff switches the tests to use a fake incrementing clock to
gaurantee timestamps are always different.

Test Plan: ran the tests

Reviewed By: k21

fb-gh-sync-id: f53d839
fbshipit-source-id: f53d839