Buck plugin handle interrupted exception
Summary:People get really scared when they see something like P56126346
This is caused by an interrupt exception which has no message.
The interrupt exception is caused by the fact that buckd dies and we stop the client.

Also, we moved the stop in the pooled thread because stop is blocking and it may slow down our UI.

So, we treat interrupt exceptions differently

Test Plan: Tested in intellij.

Reviewed By: alsutton

fb-gh-sync-id: d899916
fbshipit-source-id: d899916