Speculatively parse unflavored targets, so we can exploit parallelism
Summary:At the end of an fb4a parse, we have two targets which are explicitly fetched,
in serial, during the target graph construction. They are parsed there because
they are the unflavoured counterparts to flavoured targets in the actual graph.

This change ensures that such targets are speculatively parsed during the
normal parsing phase, in parallel with everything else.

Test Plan:Manual inspection of traces to verify that the fb4a js_bundle target does not
show up on its own at the end of the parsing phase.

Reviewed By: marcinkosiba

fb-gh-sync-id: 0ad2b42
fbshipit-source-id: 0ad2b42