Pass -fno-validate-pch to clang when using precompiled headers
Summary:This fixes a bug where mtime changes (but identical content) in PCH input files
causes clang to error with "PCH out of date" when using that PCH.

Since buck tracks the PCH input and knows when to generate a new PCH, it should
be safe to disable mtime validation in clang itself with `-fno-validate-pch`.

Note that this flag also disables some other validation like pch format
version, but since we also track the active toolchain, this should not be a
problem either.

Test Plan: Added an integraiton test.

Reviewed By: ryu2

fb-gh-sync-id: 5490186
fbshipit-source-id: 5490186