Remove broken long arg trial
Summary:The long arg test that is removed was breaking the completion for the real long arguments for all of the verbs. --version, one of the options it gave, doesn't work for any of the verbs (only for straight buck).

As with the previous diff, if you aren't the right people to review, please pass along.
Thanks!

Test Plan:1) buck build -<tab>
   See a list of long options for the build verb.
2) buck targets -<tab>
   See a list of long options for the targets verb.

Reviewed By: beefon

fb-gh-sync-id: 7c827c2
fbshipit-source-id: 7c827c2