Replace TargetGraphToActionGraph
Summary:Until now there were 2 ways to transform a TargetGraph to an ActionGraph. In the cases there was a gain from using a cache we were using `ActionGraphCache` and for the rest of the cases we used `TargetGraphToActionGraph`. This creates inconsistencies with the the events in multiple threads in `ActionGraphCache` among other things.

The refactoring ensures that there only one place that performs the transformations. If the code doesn't need caching or might slow down by checking the cache it can call static methods that don't depend on the cache but perform the same as `TargetGraphToActionGraph`. Also, most of transformation use a `DefaultTargetNodeToBuildRuleTransformer`, so there is a static method that construct it within and does not require it as an argument. The event firing is still the same and the readability of the code is somewhat improved.

Test Plan:buck autodeps; buck build buck; buck test(all ok);

One more unit test when you call the static method `ActionGraphCache.getFreshActionGraph()`, just in case the behaviour of the cached and fresh version diverge in the future.

Reviewed By: Coneko

fb-gh-sync-id: 9c151f9
fbshipit-source-id: 9c151f9