Add `--attempt-uninstall` option to instrumentation test runner
Summary:A new option for the instrumentation test runner to best effort try to
uninstall the instrumentation test package from the emulator/device. External
runners can already use the option. I need to figure out what would be a good
way of exposing the option to buck. E.g. uninstalling normal builds is useful,
but uninstalling exo builds is probably not.

Test Plan:Ran instrumentation tests with the runner, but without buck.
Verified that the instrumentation apk was gone after the test finished.

Reviewed By: Coneko

fb-gh-sync-id: ecefb38
fbshipit-source-id: ecefb38