Require absolute paths in `FileHashCache` interfaces
Summary:Previously, we'd use a mix of absolute and relative paths when using
`FileHashCache`.  In the common case of `StackedFileHashCache` if we
used a relative path, it'd work by just selecting the first inner cache
that didn't ignore the path (even if it was meant for a cache rooted
at a different path).

This diff adds checks to enforce absolute paths to prevent this ambiguity.
It would eventually be nice to have the inner caches be of a different
type and require relative paths but this will require significantly more
refactoring.

Test Plan: CI

Reviewed By: Coneko

fb-gh-sync-id: e4e7398
fbshipit-source-id: e4e7398