buildfile parser: add an "add_build_file_dep()" function
Summary:
This function can be used to explicitly indicate a dependency on an
external file.  This ensures that rules will be reprocessed correctly
when this external file changes.

In particular, this is useful if your build macros code uses an external
executable or a python archive that cannot be imported normally with
include_defs().

Test Plan: Added unittest.

Reviewed By: Coneko

fb-gh-sync-id: ad865cc
fbshipit-source-id: ad865cc