Add the ability to easily fake a buck version
Summary:When developing Buck, we often want to pretend that we haven't changed it at all relative to the version currently being used in whatever repository we're testing against. Today, that involves editing `BuckVersion.java` very carefully.

This diff allows you to add a `.fakebuckversion` file to the root of your buck repo, so pretending nothing has changed is as simple as `cp <target-repo>/.buckversion <buck-repo>/.fakebuckversion`. Naturally we warn you that this is happening.

Test Plan: Create a `.fakebuckversion`; verify that it is used as the version.

Reviewed By: marcinkosiba

fb-gh-sync-id: d256938
fbshipit-source-id: d256938