Fix Windows-only ClassUsageTracker bug
Summary:`ClassUsageTracker` was taking a shortcut and using a regex to parse the
URI coming out of `javac`. The regex did bad things on Windows, where paths
can start with things like `C:`.

The fix is to use methods on `URI` to deconstruct the URI.

Test Plan:It's really hard to test this on a Mac, so pushing to CI and letting it
run Windows builds.

Reviewed By: Coneko

fb-gh-sync-id: 090b877
fbshipit-source-id: 090b877