Enable ActionGraphCache with checking
Summary:
ActionGraphCache can save a lot of time of parsing if you build/test/install etc the same target again and again, a common pattern. This diff makes the caches utilise the cache, that means if you get a cache hit you will get the cached version.

For validation purposes we will perform a checking between the cached and a newly created ActionGraph on a separate thread with minimum priority. The checking for now will happen to 75% of the commands. The ActionGraphs will be compared using their content agnostic RuleKeys that use `RuleKeyAppendable` to `RuleKey` cache to improve the performance as much as possible and do not interfere with the build time at the same time. If there is a mismatch, it logs the BuildRules that mismatched.

If you need to disable it set `action_graph_cache_check_rate` under `[cache]` in the `.buckconfig`.
Performance improvement: For a big app the action_graph part takes from **9 sec to < 100 ms**.

Test Plan:
ant clean; buck build buck; buck test;
- Tweaked the unit tests to use a `newDirectExecutorService` for easier testing, the priority of the integrity testing is normal.
- Confirmed the generation and running of the thread. It throws properly when the check does not match.
- Confirmed `action_graph_cache_check` can disable and enforce the integrity check.

Reviewed By: marcinkosiba

fb-gh-sync-id: f3a3f6f
fbshipit-source-id: f3a3f6f