Use a special `BuckLogHandler` to handle buck logs
Summary:
Previously, we used a `FileHandler` with it's properties defedin in
the packaged logging.properties to setup logging to the buck-out/log
files.  This meant we had to hard-code the log path in both the
properties file and inside Buck.  This diff adds a custom `FileHandler`
subclass, `BuckLogHandler`, so that the log path can be determined
inside of Buck, using the same single source of truth as other code.

Test Plan: ran a build and verified logs were in correct place

Reviewed By: Coneko

fb-gh-sync-id: 8f06a68
fbshipit-source-id: 8f06a68