Correct manifest handling when rules implement `getPossibleInputSourcePaths`
Summary: When I added `SupportsDependencyFileRuleKey.getPossibleInputSourcePaths`, I changed the RuleKey calculation to make use of the new information when present, but not the Manifest creation. I discovered this when I migrated the first rule to using this support and it had cache hits after cleaning even with different source files (!).

Test Plan: Slightly tweaked an existing manifest unit test to catch this case

Reviewed By: Coneko

fb-gh-sync-id: 7f755fe
fbshipit-source-id: 7f755fe