Fix ActionGraphCache integrity check inconsistency
Summary:
The integrity check of the ActionGraphCache was getting the value of `lastActionGraph` from the class. This was creating inconsistencies when while the check thread was running you run another `build command`. Now the check thread gets the `lastActionGraph` as an argument(final) and the inconsistency is gone.

Side bonus, better logging when rules mismatch, you get the number of common rules, number of differing ones and the first 10 that are different.

Test Plan:
CI

Tested the flow that caused the issue:
- Run `buck build projectA`.
- While the integrity check for `projectA` was running I executed a `buck build projectB`.
- Run `buck build projectA` and no error message was printed and YourKit showed me the check thread finished properly.

Reviewed By: marcinkosiba

fb-gh-sync-id: b036348
fbshipit-source-id: b036348