Do not restart daemon on cache mode change
Summary:
Add some cache configuration that do not affect parsing to a set
of fields that do not cause a daemon restart. This only includes
some of the most commonly modified fields, especially `cache.mode`
which changes when `--no-cache` is passed to Buck.

Test Plan:
Run `buck build` with and without the `--no-cache` flag, check
that the flag is respected but the daemon is not restarted
between runs.

Reviewed By: Coneko

fb-gh-sync-id: b250ee2
fbshipit-source-id: b250ee2