Fix a memory leak in class-usage tracking
Summary:
We were holding on to `FileObjectTracker`, the thing that actually did the wrapping of all the objects
that the compiler uses. It remembers all the wrappers it ever gave back, so that it can give back the same ones for subsequent requests. Not freeing that up gets expensive quickly.

Test Plan: Do a large build, observe we don't die in GCs toward the end

Reviewed By: Coneko

fb-gh-sync-id: e5ec7b2
fbshipit-source-id: e5ec7b2