Make TargetNode own visibility, change visibility patterns to be first-class citizens
Summary: TargetNodes are the actual objects that can or cannot see each other in the build graph. Instead of proxying this visibility off to BuildTarget, which is ultimately just a way to reference a TargetNode, put the logic and ownership directly into TargetNode. Improve the conceptual separation between BuildTargetPattern and the new VisibilityPattern as a result.

Test Plan:
1. Attempt "buck build buck"
2. Attempt "buck build //src/com/facebook/buck/rules:"
3. Attempt "buck build //src/com/facebook/buck/rules/..."

All builds should succeed.

4. Attempt "buck build PUBLIC"

Build should fail.

5. Mark a build target as visible to each of the above patterns, including PUBLIC.

Target should successfully parse and build should succeed or fail according to appropriate visibility rules.

Reviewed By: Coneko

fb-gh-sync-id: 9c3c3a0
fbshipit-source-id: 9c3c3a0