Add non-dep-file inputs to the manifest key
Summary:
The key used for manifest file generation used to be created by using a input-based rule key builder
that had been overridden to ignore all of the `SourcePath` components of the rule key. The manifest
then contains a mapping of `SourcePath`s that were actually used (and their hashes) to the corresponding
dep-file-based rule key. When attempting a manifest-based cache hit, we'd compute the key, pull down
the appropriate manifest, and check input hashes against the current state until we found an entry for which
all of the input hashes matched what we had on disk.

However, the thing that generates the dep file often has limited purview or is only reporting certain kinds of
dependencies. For example, the C++ compiler only lists the header files that were used, not the source.

So the manifest key as currently generated ignores differences in `SourcePath` inputs that are not eligible
to appear in the dep file. For our C++ rules this didn't matter, because there's only ever a single file that
isn't eligible -- the one source file.

For Java rules there are lots of `SourcePath`s that aren't eligible to appear in the dep file -- multiple source files,
multiple resource files, a proguard configuration, resource-only dependency jars... So with the current implementation,
if you just add a source file  to a Java rule without changing any others, you can pull down a manifest with
entries for builds that were  done before the new source file existed. All of the input file hashes in one of those
entries might match your  current state, and you get a manifest-based cache hit instead of rebuilding like you should.

The fix is to include in the manifest key any `SourcePath` components that are invisible to the code that generates
the dep file. That way we're guaranteed that all of the entries in the manifest correspond to builds that differed only in
what classes actually wound up being used for the same source. The difference between manifest entries at
that point will result only from changes in the dependencies.

Test Plan: New integration test and unit test

Reviewed By: Coneko

fb-gh-sync-id: 5f3fe86
fbshipit-source-id: 5f3fe86