Move ABI classpath computation difference under the dep file kill switch
Summary:
One of the diffs leading up to turning on Java dep file support changed
how the ABI classpath was reported from reporting the archives to enumerating
all of their members. This means a lot more hashing, but given the caching
we didn't think it would make a noticeable difference.

We are seeing a big spike in one of our incremental build benchmarks that we
haven't explained yet. I'm moving this code under the kill switch in case
it's a problem after all.

Test Plan: CI

Reviewed By: nayrmi

fb-gh-sync-id: b8cd6a6
fbshipit-source-id: b8cd6a6